[chore] Move queue-batch from internal to internal/queuebatch
